+++ FathomlessVoidling/EntityStates/Utility/FireWardWipeNux.cs
@@ -153,12 +153,7 @@ namespace FathomlessVoidling.EntityStates.Utility
                 if (phasedInventorySetter)
                     phasedInventorySetter.AdvancePhase();
             }
-            foreach (TeamComponent tc in TeamComponent.GetTeamMembers(TeamIndex.Void).ToList())
-            {
-                CharacterBody cb = tc.GetComponent<CharacterBody>();
-                if (cb && cb.name == "VoidRaidCrabJointBody(Clone)")
-                    cb.SetBuffCount(RoR2Content.Buffs.HiddenInvincibility.buffIndex, 0);
-            }
+            JointThresholdController.RemoveImmunityFromAllJoints();
             FathomlessMissionController mc = FathomlessMissionController.instance;
             if (mc)
             {
@@ -168,7 +163,7 @@ namespace FathomlessVoidling.EntityStates.Utility
                 if (mc.singularityDriver)
                     mc.singularityDriver.enabled = true;
                 if (mc.mazeDriver)
-                    mc.mazeDriver.enabled = phase >= 2;
+                    mc.mazeDriver.enabled = phase >= 1;
                 if (mc.fireMissileDriver)
                     mc.fireMissileDriver.enabled = true;
                 if (mc.multibeamDriver)