+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+**0.9.14**
+
+- Improves arena lighting/visibility
+- Increases force on Gravity Bombs/Projectiles (3000 -> 5000)
+- Ward Wipe now forces gravity bombs and barnacle spawns
+- Increases Barnacle director passive credit generation
+- Wandering Singularity now lasts longer (20s -> 30s)
+- Portal Beams now properly show the indicator of where the attack will land (vanilla bug apparently)
+- Reworks joint mechanic
+    - You will now need to damage each joint down to its threshold to advance the phase
+    - Joints now have 2 thresholds, 75% HP and 50% HP
+    - Barnacles no longer spawn on the joints on thresholds
+    - After phase advance, joints are no longer immune and no longer heal
+    - Joints now cleanse debuffs when they lose 10% HP (like mithrix's dash)
+- Tweaks Eye Blast mortars
+    - Mortar count and fire delay no longer change between phases
+    - Wave count (amount of times it fires a barrage) has been reduced to 3 in P1 then increases to 5 in P2/P3
+
 **0.9.13**
 
 - Fixes joints not having adaptive armor